move compute_pgen function to utils
