+++ gen_airr_bm/data_processing/data_generation_methods.py
@@ -4,6 +4,7 @@ import numpy as np
 import pandas as pd
 
 from gen_airr_bm.core.data_generation_config import DataGenerationConfig
+from gen_airr_bm.utils.olga_utils import compute_pgen
 
 
 #TODO: This function can be rafactored
@@ -118,21 +119,6 @@ def simulate_pure_olga_sequences(number_of_sequences, model, output_file_path, s
     """
     command = ('olga-generate_sequences --' + model + ' -o ' + output_file_path + ' -n ' + str(number_of_sequences)
                + ' --seed ' + str(seed))
-    exit_code = os.system(command)
-    if exit_code != 0:
-        raise RuntimeError(f"Running olga tool failed:{command}.")
-
-
-def compute_pgen(sequences_file_path, pgens_file_path, model):
-    """
-    This function computes pgen values for the sequences in sequences_file_path and stores them in pgens_file_path.
-    :param sequences_file_path: path to the file with olga sequences
-    :param pgens_file_path: path to the file where the pgen values will be stored
-    :param model: olga model used for generating sequences (for example: humanTRB)
-    :return:
-    """
-    command = 'olga-compute_pgen --' + model + ' -i ' + sequences_file_path + ' -o ' + pgens_file_path \
-              + ' --seq_type_out aaseq --seq_in 1 --v_in 2 --j_in 3 --display_off'
     exit_code = os.system(command)
     if exit_code != 0:
         raise RuntimeError(f"Running olga tool failed:{command}.")
